Fuzzy merge approach for GIS and GAUL data
Trying to find a way to merge GIS and GAUL data based on zone names to be able to use other data files that are following GAUL zone mappings (example: food crisis data). Fuzzy merge performs reasonable but not able to match all zones, so not sure how useful this is. Maybe a starting point for other approaches though?